becalm-backend/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
openai
psycopg2-binary
django-cors-headers
gunicorn
